requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
astroid==3.2.2
asttokens==2.4.1
attrs==24.2.0
black==24.4.2
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
distro==1.9.0
docstring_parser==0.16
executing==2.0.1
fastapi==0.110.3
Flask==3.0.3
Flask-APScheduler==1.13.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icecream==2.1.3
idna==3.7
iniconfig==2.0.0
instructor==1.3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
ollama==0.3.1
ollama-instructor==0.4.1
openai==1.40.8
packaging==24.1
partial-json-parser==0.2.1.1.post4
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
promptools==0.1.3.5
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.5
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
tenacity==8.5.0
tomli==2.0.1
tomlkit==0.12.5
tqdm==4.66.5
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
